requirements.txt
feedparser
requests
PyMuPDF
pandas
openpyxl
python-docx
beautifulsoup4